==> amg_project_codes/4_BacterialClusters.Rmd <==
---
title: "R Notebook"
output:
  html_document:
    df_print: paged
---


```{r setup, include=FALSE}
library(ape)
library(phyloseq)
library(tidyverse)
library(writexl)
library(vegan)
library(microbiome)
library(ggpubr)
library(alluvial)
library(ggalluvial)
library(rstatix)
library()

setwd("~/OneDrive/University of Helsinki/msc_thesis/bact_comm")
```



```{r}
physeqR <- readRDS("PhySeq_BrakenHumGut_05.07.22.rds")
physeqR <- subset_taxa(physeqR, superkingdom %in% c("Bacteria", "Archaea"))

Pal6 <- c("#BB3E03", "#E9C46A", "#CA6702", "#0A9396",
          "#005F73", "#80BBA8")

Pal10 <- c("#9B2226","#AE2012", "#BB3E03", "#CA6702", "#EE9B00", "#E9C46A",
           "#80BBA8", "#0A9396", "#005F73", "#033549")

Pal12 <- c("#9B2226","#CC4400", "#D66915", "#E08E29", "#F0C761", "#FFFF99",
           "#C2FCFF", "#7CC6DE", "#3890BC", "#1C489A", "#000077", "#001219")

ClustersB4 <- c("#005F73","#0A9396", "#80BBA8")
ClustersB5 <- c("#005F73","#0A9396", "#80BBA8")
ClustersB7 <- c("#CA6702","#AE2012")
ClustersB9 <- c("#CA6702","#AE2012")

Clusters <- c("#005F73","#0A9396", "#80BBA8","#CA6702","#AE2012","#001219")

```



# Cluster data merging

```{r}

B4 <- read_csv("HClust_F_CLR_B4.5_18.08.22.csv") %>% 
  mutate(Cluster=paste("EC", ClusterID, sep="_")) 
B7 <- read_csv("HClust_F_CLR_B7.9_18.08.22.csv") %>% rowwise() %>%
  mutate(Cluster=paste("LC", ClusterID, sep="_")) 
P <- read_csv("HClust_F_CLR_P_13.11.23.csv") %>% rename(ClusterID = ClusterID_allWGS) %>% rowwise() %>%
  mutate(Cluster=paste("P", ClusterID, sep="_")) 

all_TP <- B4 %>% add_row(B7) %>% add_row(P)

cluster_table <- all_TP %>% separate(Sample_ID, into=c("Family_ID", "Sample_type"), sep="_", remove=FALSE)

# get info Cluster in PhySeq object
sam.new <- cluster_table %>% select(Sample_ID, Cluster)
sam.new <- sam.new %>% column_to_rownames(var="Sample_ID")

# Add new FCT in metadata 
sam.new <- sample_data(sam.new)
physeqR_data <- merge_phyloseq(physeqR, sam.new)

# add birthmode in Cluster table
sam.old <- sample_data(physeqR) %>%
    data.frame() %>% rownames_to_column(var="Sample_ID") 

cluster_table <- left_join(cluster_table, sam.old)
```

# Composition
```{r}

# composition plots
physeqR_data_compo_T  <- microbiome::transform(physeqR_data, "compositional") %>%
  subset_samples(Sample_type %in% c("B4","B5","B7", "B9","M","F"))
psG  <- aggregate_rare(physeqR_data_compo_T, level="family", detection=0.1, prevalence=0.1) 

plot_data <- psmelt(psG) %>% select(Sample, Sample_type, OTU, Abundance, Cluster)

plot_data_av <- plot_data %>% group_by(OTU, Sample_type, Cluster) %>% 
  mutate(av_Abundance=mean(Abundance)) %>%
  select(OTU, Sample_type, Cluster, av_Abundance) %>% unique()

plot_data_av %>%  ggplot(aes(fill=OTU, y=av_Abundance, x=Cluster)) + 
    geom_bar(position="stack", stat="identity") + scale_fill_manual(values = Pal10) +
  facet_grid(.~Sample_type, scales = "free_x")+
  theme_bw() +
  ggtitle("Timepoint Relative Abundance")
#ggsave("rel_abd_plot_all_samples.pdf")  
```

# heatmaps

## B4 and B5

```{r}

physeqR_data_compo_T  <- microbiome::transform(physeqR_data, "compositional") %>%
  subset_samples(Sample_type %in% c("B4", "B5"))
psG  <- aggregate_rare(physeqR_data_compo_T, level="family", detection=0.01, prevalence=0.1) 

dfm <- psmelt(psG) %>%
  filter(!OTU %in% c("Unknown", "Other"))

dfm %>% ggplot(aes(x=Sample, y=OTU)) + 
  geom_tile(aes(fill = Abundance)) +
  facet_grid(.~Cluster, scales="free_x", space="free") +
  theme_light()+
  ylab("")+
  scale_y_discrete(limits=rev)+
  #scale_fill_viridis_c(na.value = "grey50") +
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_blank()) +
  scale_fill_gradient2(low="grey50", mid="navy", high="red", 
                       midpoint=0.5)

ggsave("B4.5_heatmap_fam_13.2.24.pdf", width = 15, height = 10, units = "cm", limitsize = F)
```

## B7 and B9

```{r}

physeqR_data_compo_T  <- microbiome::transform(physeqR_data, "compositional") %>%
  subset_samples(Sample_type %in% c("B7", "B9"))
psG  <- aggregate_rare(physeqR_data_compo_T, level="family", detection=0.01, prevalence=0.1) 

dfm <- psmelt(psG) %>%
  filter(!OTU %in% c("Unknown", "Other"))

dfm %>% ggplot(aes(x=Sample, y=OTU)) + 
  geom_tile(aes(fill = Abundance)) +
  facet_grid(.~Cluster, scales="free_x", space="free") +
  theme_light()+
  ylab("")+
  scale_y_discrete(limits=rev)+
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_blank()) +
  scale_fill_gradient2(low="grey50", mid="navy", high="red", 
                       midpoint=0.5)

ggsave("B7.9_heatmap_fam.pdf_13.2.14.pdf", width = 15, height = 10, units = "cm", limitsize = F)
```
#Parental
```{r}

physeqR_data_compo_T  <- microbiome::transform(physeqR_data, "compositional") %>%
  subset_samples(Sample_type %in% c("M", "F"))
psG  <- aggregate_rare(physeqR_data_compo_T, level="family", detection=0.01, prevalence=0.1) 

dfm <- psmelt( psG) %>%
  filter(!OTU %in% c("Unknown", "Other"))

dfm %>% ggplot(aes(x=Sample, y=OTU)) + 
  geom_tile(aes(fill = Abundance)) +
  facet_grid(.~Cluster, scales="free_x", space="free",labeller = as_labeller("P_1")) +
  theme_light()+
  ylab("")+
  scale_y_discrete(limits=rev)+
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_blank())+
  scale_fill_gradient2(low="grey50", mid="navy", high="red", 
                       midpoint=0.5)
ggsave("P_heatmap_fam_13.2.24.pdf", width = 15, height = 10, units = "cm", limitsize = F)


```
## Alpha diversity

```{r}
physeq_sp <- aggregate_taxa(physeqR_data, level="species")
tab <- microbiome::alpha(physeq_sp, index = "all")
tab <- tab %>% rownames_to_column(var="Sample_ID") %>%
  separate(Sample_ID, into=c("Family_ID", "Sample_type"), sep="_", remove=FALSE)

```

# B4 & B5

```{r}

tabB4 <- tab %>% filter(Sample_type %in% c("B4","B5"))
tabB4 <- left_join(tabB4, cluster_table)

comp <- list(c("EC_1","EC_2"),c("EC_2","EC_3"),c("EC_1","EC_3"))

tabB4 %>% ggplot(aes(x=Cluster, y=diversity_shannon, color=Cluster))+
  geom_boxplot()+
  theme_light()+
  scale_color_manual(values=ClustersB4) + 
  stat_compare_means(comparisons = comp, method = "wilcox.test") + 
  ggtitle("Early Cluster Alpha Diversity")

ggsave("B4.5_Shannon.pdf", width = 10, height = 10, units = "cm", limitsize = F)

```


# B7

```{r}

tabB7 <- tab %>% filter(Sample_type %in% c("B7","B9"))
tabB7 <- left_join(tabB7, cluster_table)

comp <- list(c("LC_1","LC_2"))

tabB7 %>% ggplot(aes(x=Cluster, y=diversity_shannon, color=Cluster))+
  geom_boxplot()+
  theme_light()+
  scale_color_manual(values=ClustersB7) +
  stat_compare_means(comparisons = comp, method = "wilcox.test") +
  ggtitle("Late Cluster Alpha Diversity")

#ggsave("B7.9_Shannon.pdf", width = 10, height = 10, units = "cm", limitsize = F)

```
# Parental 

```{r}
tabP <- tab %>% filter(Sample_type %in% c("M","P"))
tabP <- left_join(tabP, cluster_table)

tabP %>% ggplot(aes(x=Cluster, y=diversity_shannon, color=Cluster))+
  geom_boxplot()+
  theme_light()+
  scale_color_manual(values="#001219") +
  ggtitle("Parental Cluster Alpha Diversity")

#ggsave("P_Shannon.pdf", width = 10, height = 10, units = "cm", limitsize = F)
```


# All

```{r}
tab_all <- left_join(tab, cluster_table)
tab_all <- tab_all %>% mutate(Time = case_when(Sample_type %in% c("B4", "B5") ~ "Early",
                                               Sample_type %in% c("B7", "B9") ~ "Late",
                                               Sample_type %in% c("M", "F") ~ "Parental"))

tab_all %>% ggplot(aes(x=Cluster, y=diversity_shannon, color=Cluster))+
  geom_boxplot()+
  theme_light()+
  scale_color_manual(values=Clusters) +
  facet_grid(.~Time, scales="free_x") +
  ggtitle("All Timpoint Cluser Alpha Diversity")
#

#ggsave("All_clusters_Shannon.pdf", width = 20, height = 15, units = "cm", limitsize = F)

```
